+++ util/opentelemetry-util-genai-emitters-splunk/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this repository are documented in this file.
 
+## Version 0.1.5 - 2026-01-27
+
+- Release 0.1.5 of splunk-otel-genai-emitters-splunk
+
 ## Version 0.1.4 - 2025-11-07
 
 - Initial 0.1.4 release of splunk-otel-genai-emitters-splunk